+++ hbw/app/javascript/packs/hbw/widget.js.jsx
@@ -21,7 +21,8 @@ modulejs.define(
         this.options = options;
         const connection = new Connection({
           path: this.options.widgetPath,
-          host: this.options.widgetHost
+          host: this.options.widgetHost,
+          payload: this.options.payload
         });
 
         this.env = {