+++ CHANGELOG.md
@@ -14,6 +14,7 @@ v1.6.0 [unreleased]
 -   [#209](https://github.com/latera/homs/pull/209) Drop react-rails & react_ujs dependencies.
 -   [#205](https://github.com/latera/homs/pull/205) Use higher order component instead of HBWSelectMixin.
 -   [#214](https://github.com/latera/homs/pull/214) Update ruby version to 2.5.3.
+-   [#213](https://github.com/latera/homs/pull/213) Add payload to widget's requests.
 
 ### Bugfixes
 -   [#194](https://github.com/latera/homs/pull/194) Set limit on the number of choices in lookup select from sql requests.